--- src/common/pages/auth/index.tsx
@@ -1,3 +1,0 @@
-export * from './Import';
-export * from './Create';
-export * from './Unlock';
