--- src/features/auth/create/index.ts
@@ -1,1 +1,0 @@
-export * from './create-identity'
