--- src/features/identity/index.ts
@@ -1,2 +1,0 @@
-export * from './identity';
-export * from './avatar';
