--- src/features/user/apps/index.tsx
@@ -1,1 +1,0 @@
-export * from './components'
