+++ src/index.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import ReactDOM from 'react-dom/client'
 import App from './App.js'
 import "@fontsource/inter";
-import './index.css'
+import './styles/global.css'
 import { registerSW } from 'virtual:pwa-register';
 
 const updateSW = registerSW({
