+++ src/stores/user.ts
@@ -1,0 +1,18 @@
+import { Identity } from '@libs/identity';
+import { create } from 'zustand';
+
+export type User = {
+  user: typeof Identity
+}
+
+export interface UserStore {
+  user: typeof Identity | undefined;
+  login: (state?: typeof Identity) => void;
+  logout: () => void;
+}
+
+export const useUser = create<UserStore>((set) => ({
+  user: undefined,
+  login: (user) => set({ user }),
+  logout: () => set({ user: undefined }),
+}))
