+++ src/views/auth/Create/SetPassword.tsx
