+++ src/views/user/index.ts
@@ -1,0 +1,2 @@
+export * from './apps'
+export * from './Dashboard'
